Default partitioning scheme implemented in defaults.py.
The installer will check an installation if there are any existing partitions, if yes, then an appropriate message is displayed for the user to choose to use the existing schema or load the default schema. Otherwise, the system is clean(no partitions), and the user is prompted to either use the default schema, or create one from scratch.