Factor out getKitComponents into kusu.util.kits
kitops.py has a method which determines which kit components are
compatible with an OS called getKitComponents. This method requires the
kit to already be present in the database. This commit factors out the
requirement for the kit to be in the database and creates
kusu.util.kits.getKitComponentsMatchingOS which is called by the
getKitComponents method in kitops.

Reviewed in KUSU-REVIEW-194.