On upgrade, check new kit's API version is valid
Upgrading kits is only to be supported with kit API version 0.4 and
newer. This patch implements the check in kitops' UpgradeAction. Along
the way, the SUPPORTED_KIT_APIS dictionary is moved from
kusu.core.database to kusu.util.kits, and kusu.util.rpmtool gained a
method compareEVR(), publicly accessible, which can be used to compare
two versions.

Work for KUSU-1264. Reviewed in KUSU-REVIEW-200.