Implement oldest_upgradeable_version check
When upgrading a kit, the new kit can specify an
oldest_upgradeable_version attribute in the kitinfo file. Kits older
than this version can not be automatically upgraded. This patch
implements this check.

Other changes include some bug fixes, such as dealing with missing
imports (compareVersion) and some adjustments to DB sessions so all
changes persist in the database.

Work for KUSU-1264. Reviewed in KUSU-REVIEW-201.