Implement "follows" logic in kit update
The kitinfo file may specify a "follows" attribute with the old name of
a component it replaces (the component is renamed). This attribute is
used to preserve component <-> nodegroup associations on upgrade in such
a case. This patch implements the logic in the kit update code to handle
the follows attribute. Other minor refactoring also included.

Work for KUSU-1264. Reviewed in KUSU-REVIEW-202.