Add -n/--remote-kid option to kitops
Currently the -i/--kid flag in kitops is used to specify a currently
installed kit for operations such as listing or removing kits, but it's
also used to specify a _remote_ kit when adding kits from a remote repo.
When updating kits, it may be desireable to identify a currently
installed kit by its ID as the kit to be upgraded, and a remote kit by
its remote ID as the new kit.

This patch adds the -n/--remote-kid (-n for "net kit") flag to kitops to
specify kit IDs in remote repositories. Code for adding kits from a
remote repo is slightly modified to use this new parameter.

Work for KUSU-1264.