Refactor remote kit code in kitops into own class
The kitops executable script contains a lot of code which handles kits
in remote repos. This commit factors all that code into its own class,
RemoteKit. No changes are made to the logic.

When adding a new kit, RemoteKit creates a temporary directory into
which files from the repo are downloaded. This directory needs to be
removed when the add operation completes. Code is added to the KitOps
class to handle this situation.

Work for KUSU-1264. Reviewed in KUSU-REVIEW-204.