Ensure kitmedia is path object in kitops
self.kitmedia.isdir() on is called on line 477 in kitops.py, but
self.kitmedia can be a plain sting which doesn't have an isdir() method.
This patch fixes the bug. Also, in setKitMedia(), assign the passed in
parameter to delete_kitmedia_on_finish instead of blindly setting it to
True each time.