kitinfo in kit RPM now contains correct filenames
The kitinfo file which was being packed into the kit RPM did not contain
valid filenames as expected. The list of filenames was empty. The
kitinfo file placed in the kit's root directory was generated *after*
the packages (and kit RPM) was built, and it contained the correct
filenames listing.

There is a chicken-and-egg problem here. The kitinfo file must include a
list of files in the kit, including the kit RPM file. The listing is
generated by listing the contents of the build artifacts directory. The
kit RPM is already generated at this point, and contains a kitinfo file
with an empty filenames list.

To get around this, we build the kit RPM twice. The first time, the kit
RPM contains a kitinfo file with an empty filenames list. Then the kit
filenames are determined in the call to populatePackagesDir(). After
this call, we rebuild the kit RPM, this time with the correct
information in its kitinfo file.

Reviewed in KUSU-REVIEW-205.