Support upgrading kits older than kit API 0.4
When a kit with a kit API older than 0.4 is upgraded, a deleteKit()
method is called according to that kit's API version. One of the
parameters passed in is a boolean indicating whether this delete action
is part of the update action. Older version of deleteKit() did not have
this parameter, and a TypeError was raised. This patch retrofits all old
addKit() and deleteKit() methods to match the signature for kit API 0.4.
The new parameter defaults to False and is never used in the body of the
method.

This patch also handles situations where the kit RPM's release differs
from that of the kit as specified in the kitinfo file, and cases where
no un/install scripts exist for the kit.

Work for KUSU-1266. Reviewed in KUSU-REVIEW-212.