First cut at implementing global locking.
Almost all kusu tools in base kit were using KusuApp so the
global lock check is added there.

The check was added to kusu-power and kusu shell (which kusu-uat
 uses) since they don't inherit from KusuApp.

kusurc also does not inherit from KusuApp but not considering
it for now since we might need to run kusurc during the native
base kit upgrade.